requirements.txt
# core
pandas
pyarrow
tqdm
python-dotenv
more-itertools

# web fetch + parsing
aiohttp
aiodns
lxml
beautifulsoup4
markdownify
readability-lxml
trafilatura
async_timeout

# document
PyPDF2
python-docx
openpyxl

# huggingface
huggingface_hub==0.*
datasets==2.*